V1.98 Fix crash formatting text on MS Windows.
git-svn-id: https://svn.code.sf.net/p/veroroute/code/trunk@252 d5f823ff-b8c7-416f-ab5d-d69718ecf442

+++ src/Board.h
@@ -441,7 +441,7 @@ public:
 	bool TakeOff(Component& comp);
 	void FloatAllComps();				// Float all components (i.e. take them off the board)
 	void PlaceFloaters();				// Try to place down all the floating components
-	void FixCorruption(size_t& nBadComps, size_t& nBadPoints);
+	void FixCorruption(int& nBadComps, int& nBadPoints);
 
 	// GUI helpers for manipulating user-selected components
 	void SelectAllComps(bool bRestrictToRects);
